Add activity for inspecting existing notifications
This should be useful for building templates for reading notifications.
The actual building of the notification table is fairly involved (a
ListView contains several TableLayouts that are generated dynamically,
but require a drawable to make cell borders), but getting the data from
the system isn't.

At one point while developing this activity the notification metadata
had to be serialized so it was packed into JSON.  After finding ways
around the serialization changing the implementation to use HashMaps
didn't recognizably improve the code, so the JSON is still around.